python_notes_server/pyproject.toml
[project]
name = "python-notes-server"
version = "0.1.0"
description = "MCP server for managing the org-roam notes"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.26.0",
    "python-dotenv>=1.2.1",
]
